# services/notification-svc/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Offense Conditions — Notification Service
  version: 0.1.0
  description: |
    Multi-channel notification delivery: in-app, email, outbound webhooks.
    Consumes domain events from Kafka, applies per-user preferences,
    delivers via channel adapters.
  contact:
    name: Offense Conditions Platform Team
    email: platform@offensiveconditions.org
servers:
  - { url: https://api.offensiveconditions.org, description: Production }
  - { url: http://localhost:8008, description: Local dev }
security:
  - BearerAuth: []
...
paths:
...
  /v1/me/webhooks:
    get:
      tags: [webhooks]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Webhook" }
    post:
      tags: [webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url, event_types]
              properties:
                name: { type: string }
                url: { type: string, format: uri }
                event_types: { type: array, items: { type: string } }
      responses:
        "201":
          description: Created — secret returned ONCE
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Webhook" }
                  - type: object
                    properties:
                      secret: { type: string }
  /v1/me/webhooks/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [webhooks]
      responses: { "200": { description: Updated } }
    delete:
      tags: [webhooks]
      responses: { "204": { description: Deleted } }
  /v1/me/webhooks/{id}/rotate-secret:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [webhooks]
      responses:
        "200":
          description: New secret
          content:
            application/json:
              schema: { type: object, properties: { secret: { type: string } } }
  /v1/me/webhooks/{id}/test:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [webhooks]
      responses: { "200": { description: Test event dispatched } }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Webhook:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        url: { type: string, format: uri }
        event_types: { type: array, items: { type: string } }
        is_active: { type: boolean }
        failure_count: { type: integer }
        last_success_at: { type: string, format: date-time, nullable: true }
        last_failure_at: { type: string, format: date-time, nullable: true }
        last_failure_msg: { type: string, nullable: true }
        disabled_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...